bootstat: fix RecordInitBootTimeProp was not sending a metric
The time unit of `ro.boottime.init` is nanoseconds, and its size is 64
bit. Since its values are nanoseconds, they easily become larger than
the max of a 32 bit integer. However, RecordInitBootTimeProp parses them
as Int32, and thus fails to read them. As a result, bootstat is not
sending values of `ro.boottime.init` at all today.

This commit fixes this issue by parsing values as 64 bit integers.

Other properties in the `boottime` family, such as
`boottime.init.cold_boot_wait` and
`ro.boottime.init.(first_stage|selinux)` have the same issue, but they
don't exceed the max of 32 bit integers because they are duration, not
timestamps. They have a different issue: their time unit is ns when it
should be ms. However, this issue is out of the scope of this commit.

Bug: 430701171
Test: Confirmed now `ro.boottime.init` is being sent
Flag: EXEMPT BUG_FIX

+++ bootstat/bootstat.cpp
@@ -1186,7 +1186,7 @@ std::string_view CalculateBootCompletePrefix() {
 void RecordInitBootTimeProp(BootEventRecordStore* boot_event_store, const std::string& property) {
   auto value = android::base::GetProperty(property, "");
 
-  int32_t time_in_ms;
+  int64_t time_in_ms;
   if (!android::base::ParseInt(value, &time_in_ms)) {
     return;
   }
@@ -1197,7 +1197,7 @@ void RecordInitBootTimeProp(BootEventRecordStore* boot_event_store, const std::s
     time_in_ms /= 1000000;
   }
 
-  boot_event_store->AddBootEventWithValue(property, time_in_ms);
+  boot_event_store->AddBootEventWithValue(property, static_cast<int32_t>(time_in_ms));
 }
 
 // A map from bootloader timing stage to the time that stage took during boot.
